requirements.txt
# Cloud SDKs
boto3>=1.28.0
google-cloud-storage>=2.10.0
google-cloud-bigquery>=3.11.0

# Orchestration
apache-airflow>=2.7.0

# Big Data
pyspark>=3.4.0

# Data Quality
great-expectations>=0.17.0

# Utilities
pandas>=2.0.0
numpy>=1.24.0

# Testing
pytest>=7.3.0
